GML: fix MSVC warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk@31608 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/gml/gmlpropertydefn.cpp
@@ -175,7 +175,7 @@ void GMLPropertyDefn::AnalysePropertyValue( const GMLProperty* psGMLProperty,
             {
                 /* grow the Width to the length of the string passed in */
                 int nWidth;
-                nWidth = strlen(pszValue);
+                nWidth = static_cast<int>(strlen(pszValue));
                 if ( m_nWidth < nWidth ) 
                     SetWidth( nWidth );
             }
